refactor shadow condition and judgement engine
