+++ shadow-core/shadow-core-rewrite/src/test/java/org/apache/shardingsphere/shadow/rewrite/PreparedJudgementEngineTest.java
@@ -19,6 +19,7 @@ package org.apache.shardingsphere.shadow.rewrite;
 
 import org.apache.shardingsphere.api.config.shadow.ShadowRuleConfiguration;
 import org.apache.shardingsphere.core.rule.ShadowRule;
+import org.apache.shardingsphere.shadow.rewrite.judgement.impl.PreparedJudgementEngine;
 import org.apache.shardingsphere.sql.parser.relation.metadata.RelationMetas;
 import org.apache.shardingsphere.sql.parser.relation.statement.impl.InsertSQLStatementContext;
 import org.apache.shardingsphere.sql.parser.sql.segment.dml.column.ColumnSegment;
